Added the feature where if you specify --flatten_linker_scripts in the command line arguments then the CompilerStub will
1 - Remove --flatten_linker_scripts from the arguments list.

2 - For each .o file, check if it starts with the string "INPUT".  If
so it will parse it as a list of files.  The original .o file is
removed from the command line and replaced with this file list.  These
filenames are then also checked, so it implements a Breadth-First
Search on the filename graph.

There is verbose debugging output which can be turned on by setting
the $linker_script_debug variable in the script.