Removed these lines in the help message.
  --ppargs=xxx An argument to be passed to the preprocessor (can be specified
               multiple times)

At the end of the help message it says as follows.

  All of the other arguments are passed to the preprocessor.

Which is in fact what happens to *any* arguments passed in as --ppargs
as well.  That is --ppargs is not recognized at all, and is also
redundant.

For example, if you call patcher with this flag

  --ppargs=-I/gronk/foo

the preprocessor gets the whole --ppargs=-I/gronk/foo, not just the
intended -I/gronk/foo .