This fixes (I hope!) the Scott/Dan problem with tagged wild functions and separate compilation.
Loosely, the problem is as follows: In one file you have
  int foo(int) __WILD {body}
and in another you have
  extern int foo(int) __WILD;
  int (*fptr)(void) = foo; // must be tagged!
  fptr(5); // call WILD TAGGED function
and in a list of functions to be tagged you have:
  foo

So we have WILD functions and we want them tagged. Changes:

(1) We were creating "foo__descriptor" in the one file but looking for
"foo_tw__descriptor" in the other. I now name all function descriptors with
the pre-mangled name (e.g., "foo__descriptor"). Since only WILD functions
can have descriptors, this should be no problem.

(2) We weren't producing descriptors for functions that were declared in a
file but never used (the production step was associated with encountering
the function's variable somewhere). We do now.

(3) There was some confusion over "foo_w" versus "foo_tw". Since tagged
functions are always WILD *and* suffices are supposed to prevent link
errors *and* tagged and untagged WILD functions link the same way (since
the check, if any, should happen at the call site) there is no reason to
distinguish between them, so tagged wild functions are now named simply
"fun_w". The link error (if the callee is not tagged) will show up as a
missing descriptor.

It is entirely possible that some key assumption here is wrong. This fix
allows their small test case to proceed, however.