When printing expressions, put parens around arithmetic comparisons that appear as operands to bitwise operators.  For example, print this:
	(a == b) & c

instead of this:

	a == b & c

Comparisons bind more tightly than bit twiddlers, so both forms are
semantically equivalent.  However, the non-parenthesized version
elicits a warning from "gcc -Wall" suggesting that parentheses be
added.  We shouldn't be generating code that elicits warning
diagnostics not present in the original C code.

(There was already some logic present for a similar situation with "+"
and "-" as operands of bitwise operators, which has been reworked
somewhat and folded in along with my changes for comparisons.)