Add a small test which shows that CIL produces slightly bogus code for initializing const arrays.
I don't think this will actually yield a failure when regression
testing, though, because gcc only prints a warning rather than an
error.  Presumably we want $(WARNINGS_ARE_ERRORS) to be turned on in
test/Makefile when building this.  I don't know how to do that,
though.  If someone else does, please make the necessary change and
check it in.  Thanks!