Update the epicenter code to reflect changes in the Rmtmps API. Although Scott tells me that this code is no longer used, it does serve as a nice example of using Rmtmps with a custom global roots marker.
(Actually, it's the *only* example of using a custom global roots
marker.  If the epicenter code were not here, there wouldn't really be
any clear reason to support that functionality in Rmtmps in the first
place.  This functionality is cheap enough to provide that I don't
mind having it in there, but if someone eventually removes
"epicenter.ml" then it might make sense to simplify Rmtmps's API
accordingly.)

The call graph module has been modified to store more detailed
information in each graph node.  Previously, we recorded only the name
of the function described by that node.  Now re record the full
varinfo instead.  The name is still available from the "varinfo.vname"
field, but this way we also have access to other information about the
function.  In particular, the epicenter code can now set
"varinfo.vreferenced" bits as it walks through the call graph.