Two new regression tests in the "cil" group:
Test "small1/warnings-empty.c" shows that even the most trivial
possible C program comes through CIL with new warnings that did not
exist in the original.

Test "small1/warnings-noreturn.c" shows that CIL adds an explicit
"return" statement to the end of each function, which yields a warning
when that function was declared "noreturn".