In the Perl "Cilly" compiler driver base class, add a "setDefaultArguments" method which is called just before command line argument processing.
The default method does nothing, but this provides a hook for
subclasses to inject their own code to set default values which
subsequent command line flags might override.  Previously, one had to
override the "new" method and install each default value only if it
was still undefined *after* command line processing, which is
decidedly less convenient.

Use this in "CCured" subclass to set default cure type to "infer".
Set MERGE_INOBJ to 1 as well, since that was being done previously.
I'm not exactly sure what this is used for, though.