Only echo the commands being executed if $self->{TRACE_COMMANDS} is true.  This defaults to true, and there is no command line flag to turn it off.  So the default behavior will be the same.  However, one can change the default to false in subclasses by overriding Cilly::setDefaultArguments().
Turning off command tracing is desirable when trying to create a tool
that pretends to be a regular compiler.  Specifically, certain
autoconf tests expect the compiler to produce *no* diagnostic output
at all unless there was an error.

The sane thing to do would really be to condition command tracing on
$self->{VERBOSE}, which is tied in to the "-v" or "--verbose" command
line flags.  This is, in fact, how it used to work.  However, Scott
explicitly disabled that because "there's already so much noise in the
output, and this is the *one* piece of information I *always* end up
digging around for".  I don't want to simply revert Scott's change, as
he clearly had a good reason to make it.  But for my work, I do need
some facility for silencing the driver.