Revamp Perl wrapper scripts' strategy for dealing with command lines. Previously, each command line was assembled as a single string.  When this reached Cilly::runShell() and ultimately Perl's system() function, that single string would be split back out into an argument vector by the underlying shell.  This is a brittle approach that fails spectacularly if there are any special characters in the command line. Cilly::fetchEscapedArg() made a small effort to quote whitespace, but did nothing about star, backtick, semicolon, etc.
The new strategy always treats a command line as a vector of strings.
That's how they are assembled internally, that's how they are passed
to Cilly::runShell(), and that's how they are ultimately given to
Perl's system() function.  When the latter is called with an array
rather than a single string, it uses the array directly rather than
going through the host's shell.  The result is much improved
robustness in the face of non-alphanumeric characters, such as
whitespace or any punctuation you care to toss in.