Reworked management of temporary/intermediate output files in Perl driver scripts.  Default behavior is now to remove all such files, just like any regular production compiler would do.  To keep intermediate files, use one of the following existing flags:
    "--keepmerged" for the single-file merge of all sources

    "--keep=<dir>" for various other CIL and CCured output files

    "-save-temps" for various gcc intermediate files;
    		  MSVC has no equivalent option

Yes, it is confusing having three different flags.  But George has
specifically requested that there be a way to retain the CIL/CCured
output without also keeping gcc's intermediate files around.  All
three of these flags already existed, and I've kept the meaning of
each as close as practical to what it was before this change.

As part of this change, some intermediate files have changed their
names slightly so that new suffixes are always preceded by a period.
For example, CCured output that used to appear in "foocured.c" now
appears in "foo.cured.c".

Because we are creating names for output files, and because these
names are often derived from the names of input files with modified
suffixes, there is some risk that a bug could cause an input file to
be overwritten.  I've done more than the usual amount of testing, and
I've also put in some safety interlocks that *should* prevent any
input file name from ever being selected for output.  However, I am
fallible and therefore there is still some risk.  Please be careful
for the first few days after picking up this change.  It would not be
excessive to make backup copies of all irreplaceable source files in
any given directory before running the CCured or CIL driver scripts.
Better safe than sorry.