Reimplement internal utility routine Cil.escape_char.
The old version used a rather baroque means of deciding whether a
character should be octal escaped, and an even more baroque way of
generating the ocal sequence.  The new version is simpler and avoids
considerable wasted effort in the common case of printable characters
that do not need to be escaped.

The old version treated many (perhaps all) characters in the 128
.. 255 range as though they were printable, which will only work if
the Caml runtime and the C compiler both agree on how to interpret
them.  The new version plays it safe and octal escapes all such
characters.

The old version escaped NUL using only a single octal digit: "\0".
This is incorrect in any string containing NUL (ASCII 0) followed by
any character '0' .. '7'.  The new version consistenly uses three
octal digits when escaping, including escaped NUL, and therefore
avoids this bug.

The old version used alphabetic escape sequences for only the subset
of sequences shared in common between C and Caml.  The new version
uses alphabetic escape sequences whenever possible, including several
which are present in C but not Caml.