Simplified the API for customizing the root set used by Rmtmps.removeUnusedTemps.  Previously, clients were supposed to set various "referenced" fields directly.  Now, clients provide a test predicate which will be invoked on each global.  The predicate should return true if the given global is to be treated as a root.
Besides being a simpler API, this also yields correct behavior in some
cases where the old approach failed.  The problem with the old
approach was that setting some "referenced" bit did not implicitly
traverse into other supporting symbols that the kept item might need.
For example, marking a global variable as referenced did not mark
necessarily mark that variable's type as referenced.  The new approach
reserves referenced bits for use by the reachable items marker, and
manages those bits in a consistent manner so as to avoid this problem.

Clients that need custom root sets should *not* set "referenced"
fields directly; the only approved way to extend the root set is via
the optional test predicate to Rmtmps.removeUnusedTemps.