Do not silently pad out function calls with zeros for missing arguments.  Instead, if the function is called with too few arguments, just leave it that way.
CIL's habit of silently turning invalid programs into ones that are at
least syntactically correct can create some subtle bugs later on.  In
this particular case, Evolution's configure script was trying to
figure out whether gethostbyname_r() took five arguments or six.  It
compiled a five-argument call to see if any errors were reported.  An
error does appear when using regular gcc, but that error vanishes if
you're using a CIL-based front end that silently pads out missing
arguments.  The result is that Evolution incorrectly thinks it can do
a five-argument call to gethostbyname_r(), and crashes at runtime when
it tries to do so.

If some CIL-based analysis actually needs to have these extra
arguments added in, it should be done in a distinct module that the
client analysis calls explicitly but which is *not* part of the
standard processing chain.