More work on the difference between simple and extended "asm" directives.  All of this applies only to GCC, of course.
Teach the CABS stage that these are two different things, and remember
which one we have by changing the Cabs.ASM structure.

When converting from CABS to CIL, though, erase this difference.
Rewrite simple directives as extended ones.  This requires escaping
"%" by rewriting it as "%%".  After that rewrite, it should be safe
for CIL to always print these in extended form.

Update regression test accordingly.  Add directives in both formats,
and use the preprocessor to reduce the test to nothing unless using
GCC on x86.