Type attributes like "const" are downward contagious across field access and array index operations.  In other words, if "x" is a const-qualified struct, union, or array, then "x.f" or "x[n]" is const as well.  I'm surprised nobody bumped up against this before.
Added a regression test which uses explicit casts to remove this
const-ness before assigning.  Earlier versions of CIL considered the
cast redundant and removed it, leading to a transformed C file that
would not compile.  With this change, CIL understands that the source
and destination types in the cast are different: the former has const;
the latter does not.  It retains the casts and therefore the
assignments compile.