Disable the printing of array length attributes as /*[n]*/.  When the pType method uses pAttr to print attributes, it sometimes decides that these attributes should be commented out.  If the attribute itself is printed as a comment, then you get nested comments.  C does not allow comment nesting, so things fall apart pretty badly.
I've added a test, "test/small1/array-args.c" that is supposed to
demonstrate the problem if you enable printing of array length
attributes as comments.  Unfortunately, the problem does not actually
appear when run using CIL's regression test infrastructure.  I think
this is because regression tests are run with the global
print_CIL_Input flag turned on.  When print_CIL_Input is on, the outer
set of attribute-disabling comment markers is omitted, so the comment
nesting problem never appears.  So I'm not sure how to get this
regression test rigged up properly.