Add --no-convert-field-offset to dosimplify
By default, if --dosimplify is used, field offsets (expressions of the
form "&a->f") are converted to dereference of a's address plus offset in
bytes.  The new option suppresses such conversion, and keeps other
simplifications.

The aim of this option is to make --dosimplify-generated code better
verifiable by static code checking tools.