Add a default context for global scope
The attached program, test/small1/scope12.c, compiles with gcc, but not cilly.
In a function scope, a new type symbol is introduced. Then a global variable is
introduced with the same name. Later, CIL detects references to the variable as
references to the type instead.

I managed to get CIL to proceed by adding a default context (for global
variables), but this solution is not well tested.

Closes #3532283.