Fixed bug in which pointer difference operations were incorrectly typed.
A subtraction between pointers was previously incorrectly considered to have
the type int. The correct type is ptrdiff_t, which may not be the same as int.
Fixed by adding a new type reference variable ptrdiffType that is set on
initialization in a way similar to upointType and the fixed the cabs2cil
translation to construct the pointer difference with the correct type.

Without this fix the resulting tree could break the invariant that both
arguments of a binary operator will have the same type (casts are inserted
to make this so. For example if "p" and "q" are pointers and "i" is an int
then CIL would previously have allowed the expression "(p - q) + i", which
violates this invariants betcause the two arguments of the "+" are of
different types. With the fix CIL inserts the necessary cast to "i".