Fix MSVC build parameters for MSys
Invocations of MSVC cl compiler were using "/" as the option
switch rather than "-". This isn't needed - cl understands
"-" as well as "/". Not only is it not needed, but it is
not wanted because when invoking cl through the MSys bash
shell MSys will see the "/" and attempt to treat what
follows as a Unix style path that it will "kindly"
translate to a Windows style path before calling cl.

There are more instances of this issue in Perl scripts, but this
is enough to get Cil built, and I don't speak Perl.