Offer choice between Olf and Golf in Ptranal
Default to Golf analysis, and provide --ptr_use_olf flag for users
prefering Olf (Golf should always yield better results).

We use OCaml 3.12 first-class modules to switch the implementation
based on a command-line flag. Because of poor design choices in the
API of ptranal.ml, it is unfortunately not possible to switch dynamically
between both analyses: the choice has to be made only once, when the
module is loaded.

Closes #146.