Add ability to have size_t on 8 bytes
Under mingw on x86_64, size_t is represented with 8 bytes/64bits, but long
is represented with 4 bytes (under Windows, long is always 4 bytes, under
Linux, long is 8 bytes on x86_64, which is why CIL could be configured on
Linux x86_64 but not Windows x86_64).
We add the option to compare the size_t type with "long long" integers.