Avoid warnings about casting pointers to int
intptr_t is defined to be an integer type large enough to hold any
pointer value, so that's what we should cast to if it is available.
These specific values are just alignments, though, so it's fine for
those intptr_t values to subsequently be coerced to int.  That also
simplifies printing, since it means we don't need wacky macros like
PRIdPTR to build printf format strings.