Fix a bug where initializations of global variables were not visited
When merging CIL AST, it may happen that the declaration of a global
variable is seen before its definition. Because of that, the vinit part of
the variable is empty and the init part of the GVar may contain something.
But the CIL visitors never look at the GVar init and only explore the vinit
part of the variable.
This caused the merging to:
- Break the v.vinit = init of the GVar invariant
- Not unify the vid of the variables used in the init part of the GVar.