Fix renaming bug triggered by static locals hiding other locals/formals.
When a static local is promoted to a global, the alpha renaming usually
remembers any suffix added and chooses a fresh suffix for later vars of
the same name. However, if the same function also has a (non-static)
local with the same name, this gets clobbered: the alpha renamer throws
away the renaming state after the function scope is done. This can be
triggered if (among other cases) two functions in the same file both
have like-named statics hiding other locals/formals (seen in Motif).