Special treatment of __builtin_classify_type
Add special treatment of __builtin_classify_type to Cabs, just like for example for __typeof__.
Turn into CLASSIFYTYPE, and in cabs2cil.ml evaluate it to the appropriate int constant that GCC also uses.

This is necessary to be able to determine the correct type of the conditional expressions used for tgmath.h.
The correct type for an expression such as
0 ? (float *) 0 :  (void*) (__builtin_classify_type(...) == X) depends on whether the last argument is a null pointer constant.
`(void*) (__builtin_classify_type(...) == X)` is a nullptr constant iff `__builtin_classify_type(...) == X` is zero.
To determine this it is neccessary to evaluate __bulitin_classify_type(...).