heapify: Include <stdlib.h> in generated C code
It is required for declarations of the malloc and free functions,
otherwise future compilers will not be able to build the generated
code because they do not support implicit function declarations by
default.

This also addresses potential crashes on 64-bit architectures if
the 64-bit pointer returned by malloc is truncated to 32 bits
because of the int return type of the implicit malloc declaration.